debug(laser af): only catch Exception, print exec_info, indicate failure for caller in autofocus method (#75)
This just adds in more debugging for the case when auto focus fails, and
gives an explicit printout in our acquisition loop if autofocus fails
(and makes it clear that we just continue anyway!).

This just adds debugging, and shouldn't affect actual behavior.

Also a few logic cleanups to quiet the ide linter.

Tested by: I ran this on a system with autofocus. To get autofocus to
fail, I had it scan onto a plate edge where it wouldn't find the 2 laser
autofocus dots and made sure that the scan continued (but that we got
printouts).