fix(acquisition abort): Aborting acquisitions midway left the ui in an invalid state (#76)
`34713ec14805b441a1cee7dbeb2c13c337035d31` added the
`self.is_current_acquisition_widget` concept on our acquisition widgets.
This was set to `= False` in two places for each - one in the
`toggle_acquisition` and another in `acquisition_finished`. The
`toggle_acquisition` always ran first (it sent the signal that
eventually resulted in `acquisition_finished` getting called), which
means that the code in `acquisition_finished` was always skipped via the
`if` guard checking for `is_current_acquisition_widget`.

It's not clear if we need `is_current_acquisition_widget` since it's
new-ish, but for now leave it and just make it so `acquisition_finished`
is the only `= False` setter of that flag.

Tested by: running local acquisitions. Before this, the stage control +
acquisition widgets would get wedged in disabled state if you aborted an
acquisition midway. Now, they don't and you can continue using the UI.