acquisitions: catch TimeoutErrors in acquisitions, and abort acquisition cleanly (#81)
Reported by @Alpaca233 .

This is not the long term solution - we should be able to handle single
timeouts (when a retry succeeds). But we do need some top level
mechanism to make sure the gui doesn't wedge in a useless state.

Tested by: Tests pass, and a local sim with a `raise TimeoutError()` in
`def run` behaves as expected.