utils: add a directory creation helper, and use it instad of os.mkdirs (#82)
This fixes the case of selecting a save path that does not exist. As
long as you have permissions to create all the directories leading to
your chosen path, they will be created for you.

Tested by: Setting my save path to `/home/ian/does/not/exist/Downloads`
and making sure acquisitions succeed and put the results there. Before
this change, acquisition failed with that save path (which does not
exist). Unit tests also still pass.