feat(microcontroller): Add a mechanism for automatic serial reconnect, and use it (#78)
We know our long running acquisitions can sometimes hang. We think this
is, at least sometimes, due to serial hiccups. This adds a mechanism for
automatic reconnect of our `Microcontroller` serial device. For now, it
does not add this same functionality to all our `serial_peripheral`
devices, but doing so should be easy.

This does not handle the cases of a caller specifying a timeout, and the
timeout elapsing. In those cases, the caller is responsible for catching
the `TimeoutError` and doing something about it.

Tested by: Running a local sim, and running acquisitions on a real
system. When running on the real system, I unplugged and plugged the
microcontroller USB a bunch during acquisitions. The only time I got it
to not resume was when I got an "mcu operation timed out" TimeoutError
to get thrown, which is expected - these are the cases when a timeout is
specified, it elapses, and we let the caller know (so the caller needs
to handle that case).