acquistions: when we abort, do not let another start until previous is finished (#96)
This was a bug whereby if you aborted, and then spammed the start
button, you could cause the ui to crash.

This stops this from happening by checking to see if the worker thread
is still running. If it is, we don't allow another to start.

Tested by: running on the Prior system.