fix: handle piezo presence/absence (#124)
We weren't handling the `False` case for piezo presence. This fixes
that, and adds a basic test to make sure we can instantiate the high
content screening gui with or without the piezo.

Tested by: Added a unit test to make sure we can at least create the gui
with or without a piezo.