fix(logging): Make sure our file handler uses unicode and continues through errors (#130)
This is a reported bug from a windows user. The logging system was
tripped up by the lower case mu character.

I don't have a system that can replicate the issue, but this seems
likely since there are some documented cases of windows defaulting to
ascii.

Tested By: I cannot test this locally. There used to be a
`sys.setdefaultencoding` hack, but it sounds like the default encoding
is hard coded past python 3.8 (see
[here](https://stackoverflow.com/questions/2276200/changing-default-encoding-of-python).
I could force FileHandler to use ascii, but that doesn't really seem
like an actual test...