fix(tests): microscope init test fix, other argument change fixes (#136)
Most of the fixes here are just changed object structures, not actual
bugs. But the `microscope.py` initialization issues are real and would
prevent `microscope.py` from working as expected.

Tested by: More unit tests pass now. There's still 1 test that segfaults
on a `QApplication.processEvents()` call, but not because of changes
here.