cameras: start plumbing AbstractCamera throughout (#111)
This starts working through the code and changing all the camera uses to
be conformant to the `AbstractCamera` interface. As a result, it depends
on #61 .

After this is finished, one more PR will need to happen that changes the
`Camera` implementations to be `AbstractCamera` conformant. I'm doing
this one first so we can get everything running with the simulated cam,
then pick off the camera implementations 1 by 1.

Major Remaining Items:
* Decide how to do the callback enable/disable. It isn't great that one
part of the code can disable callbacks (and then forget to re-enable
them, or faile to because of an exception) and bring down everything
else that cares about images. To get around this we added the callback
registration mechanism, but as a result we don't have a good way for
non-relevant images to be ignored. I haven't decided what to do for
this, yet.

Tested By: Unit tests and lots of on hardware testing.  There are 2 small follow up fixes coming as a part of this merge group.