cameras: add the default gxipy camera implementation (#145)
This adds the camera implementation we fall back to when no special
camera driver is specified. As part of adding this, a few camera configs
needed to be added (eg: default white balance).

Also, some of the functionality, like camera temperature, is not
supported via this interface. Right now, these raise
`NotImplementedError`.

Tested by: This needs testing on actual hardware. I don't have a camera
I can use it with right now. The gui does start and run on HW using
other cameras, but we need to test with one of these cameras.