camera: simulated: fix continuous streaming mode, exposure, other (#187)
This fixes the simulated camera such that it replicates the behavior of
the other cameras: you need to send a trigger to get frames, unless in
continuous streaming mode.

It also fixes the continuous streaming mode for the sim cam which was
broken before, and caused printouts even when not in continuous
streaming mode.

NOTE: There's still an issue whereby we do not turn off streaming after
acquisitions, or properly turn off streaming after live mode. So
sometimes if you're in continuous mode, you'll see that frames are still
streamed (because we don't turn off streaming). This is a separate issue
that we need to fix.

Tested By: Unit tests pass, also when in simulated gui mode no printouts
from the streaming thread occur unless in continuous triggering mode.