tests: skip debug plot test because backend (#198)
We get errors when running this as a github action because qt is
pre-loaded, and it conflicts with tk. It isn't critical, so to get the
tests running skip it for now.

Also, now that tests are running it revealed the width,height swap in
the simulated camera `_next_frame`. This also fixes that.

Tested By: Workflow must pass!