tests: sim cam: force no flip or rotate (#202)
If the config this test grabbed had rotation or flipping, we can end up
with an issue where the resolution doesn't match the resulting frame, so
the test fails.

This actually highlights a potential api issue - it's surprising when
resolution does not match the frame dimensions. But since we allow
cropping, rotating, flipping, etc it'd get really complicated to carry
that through to resolution (and it's not clear that's what you'd want
anyway).

For now, resolution is what comes off the camera. And there's no
guarantee the frame will match that.

Tested by: Tests pass with rotation angle = 90 now.