microcontroller: ignored packets with bad checksums, scan input buffer (#196)
This makes sure the incoming crc are correct, and if not scans along the
read buffer trying to find a valid message border. It also changes our
logic to get rid of the exact packet length match in favor of just
requiring enough bytes for a valid packet.

Tested By: Running `tools/microcontroller_stress_test.py` for a while
with #195 on the firmware side. I also tested without #195 on the
firmware side to make sure that case still works.