Microcontroller: Add a read-staleness watchdog (sortof) so we can see if new data is coming in (#199)
This will tell us if new data is still coming in from the
microcontroller. If we don't see printouts, valid packets are coming in.

Tested by: I set the stale timeout to <10ms to make sure this prints
out. Also, I ran the stress test with `--no_wait` which locks up
responses on the microcontroller. Both of these made prints occur.