microscope: Eliminate 1 move in z backlash compensation (#190)
We noticed that our z move times are slow, and we need to speed them up.
This helps by minimizing the number of moves needed for backlash
compensation (it should never be more than 2 moves total for a z axis
move, and in 1 direction it should be 1).

Tested by: Tests pass, local sim works, and running moves on hardware
works.