z movement: move backlash compensation into CephlaStage so it is universal (#207)
Before, we'd do z backlash compensation in one-off and sortof
inconsistent ways. Also, the `PriorStage` has its own internal backlash
compensation, so in that case we'd be doubling up on backlash
compensation.

This changes our convention to always do backlash compensation in the
stage implementation, so users do not need to care (and always get
backlash compensation if doing blocking calls).

The next step is to move backlash compensation down to the embedded
level, but I looked into doing that today and it was not
straightforward.

Tested by:
1. Running a small acquisition
2. Run the stage timing script a bunch for z axis moves.