core: MultiPointWorker: separate into its own file (#223)
This is a pure refactor / move PR in preparation for work on separating
the ui and MultiPointWorker.

Tested by: Running sim mode, running the unit tests, making sure no
errors show in pycharm, and running a bunch of acquisitions on a system (we didn't have laser AF or hw trigger, however).