MultiPointWorker: Use callbacks for image processing so we can trigger faster (#242)
This is step 1 of 2 to get to full overlapped triggering acquisitions.
This adds in the callback mechanism in `MultiPointWorker` for handling
frames asynchronously, but only triggers one frame at a time (and uses
the `threading.Event()` flag to note when the frame comes in). This is
still faster than what we had before, because we can move and do other
processing after the full frame time passes, but is still slower than
doing overlapped triggering.

The next step after this works will be to add overlapped triggering.
That is a bit tricky because we need to be 100% sure we don't drop any
frames, or else we can do an entire acquisition with all the remaining
frames (after the dropped one(s)) shifted!

Tested by: I ran a local simulation and unit tests pass, but this needs
actual testing on hardware!