MultiPointWorker: Use trigger-time and move-time positions throughout.  Do not recapture. (#250)
This is important for the ui since we can now be moving significantly at
some points when we used to not be. In general, we don't want to use
`get_pos()` on the stage outside of the real moments when we care about
the current position.

Tested by: Running acquisition on a squid.