bug fix: Prior stage config update; Daheng camera model support (#259)
This pull request includes updates to configuration files and
camera-related code to enhance functionality and ensure compatibility
with additional hardware. The most important changes involve adjustments
to homing and position limits in the configuration file, as well as
support for a new camera model.

### Configuration Updates:

*
[`software/configurations/configuration_Squid+_H117_ORCA.ini`](diffhunk://#diff-60d10781037ed93624c6ffe737893efe804d2cc1ed2a50885c749610a4f143e2L113-R113):
Changed `homing_enabled_z` from `True` to `False`, and updated software
position limits for X, Y, and Z axes to refine movement boundaries.
[[1]](diffhunk://#diff-60d10781037ed93624c6ffe737893efe804d2cc1ed2a50885c749610a4f143e2L113-R113)
[[2]](diffhunk://#diff-60d10781037ed93624c6ffe737893efe804d2cc1ed2a50885c749610a4f143e2L222-R227)

### Camera Compatibility:

*
[`software/control/camera.py`](diffhunk://#diff-1f0ea86f2716d10c7fac39432a662c1ba9cb884c850a7f46602fbbc8805473f4R330):
Added support for the `MER2-1220-32U3C` camera model by mapping it to
the `IMX226` sensor.
*
[`software/squid/config.py`](diffhunk://#diff-d721a332424b0a4636d89fa4bf1d5eb87817e778a9073dc2731cde5dcbfd8473R174):
Extended the `GxipyCameraModel` enum to include the `MER2-1220-32U3C`
camera model for compatibility with new hardware.