home x first to make sure stage is clear for homing (#265)
As we removed the moving stage step before shutting down, the current
clearing strategy may result a collision of the plate clamp post and
transillumination arm. To fix this, home x first before the +20 move, so
x will be somewhere in between the loading position and the
transillumination arm.
 
Tested on a Cephla stage.