line profiler - keep manual range if manual range is specified (#270)
This pull request introduces enhancements to the line profiler
functionality in `software/control/core/core.py`, including the addition
of a flag to track manual y-range adjustments and logic to prevent
auto-scaling when the range is manually set.

### Line profiler functionality enhancements:

*
[`software/control/core/core.py`](diffhunk://#diff-2fc90167f3cd3e8a6b36bdd78086c6ae5811054d9f87239fa2fe2b3a8858d691R3038):
Added the `line_profiler_manual_range` flag in the `__init__` method to
track whether the y-range of the line profiler plot is manually
adjusted.
*
[`software/control/core/core.py`](diffhunk://#diff-2fc90167f3cd3e8a6b36bdd78086c6ae5811054d9f87239fa2fe2b3a8858d691R3143-R3149):
Connected the `sigRangeChanged` signal of the `line_profiler_plot` to a
new `_on_range_changed` method in `toggle_line_profiler`. This method
sets the `line_profiler_manual_range` flag to `True` when the range is
manually changed.
*
[`software/control/core/core.py`](diffhunk://#diff-2fc90167f3cd3e8a6b36bdd78086c6ae5811054d9f87239fa2fe2b3a8858d691L3224-R3233):
Updated the `update_line_profile` method to conditionally auto-range the
plot only if the `line_profiler_manual_range` flag is not set.