gui: use the primary screen for initial window size (#279)
Before, we'd calculate this for the entire display area. If a user was
using a multi screen setup that had monitors that didn't line up
perfectly, they could get an unusable window. This now only calculates
the window size based on the primary display size.

Tested by: Running on a system with multiple monitors and making sure
the main window is now usable.