cameras: toupcam: fix timing to respect camera "trigger delay" and fix hardware triggering (#282)
This is a two fold change. First, the toupcam has a "trigger delay" that
depends on the exposure time, capture dimensions, etc. We weren't
calculating and using this before. The attached diagram outlines the
issue.

So, with this, we now:
For Software Triggering:
1. Send the camera the raw exposure time in ms for its exposure_time.
2. Turn on illumination right before we send the software trigger.
3. Turn off illumination after trigger_delay + strobe_time +
exposure_time.

This results in all rows getting exposure_time of illumination while
capturing data.

For Hardware Triggering:
1. Send the camera exposure_time + strobe_delay for exposure time
2. Use exposure_time + strobe_delay to calculate trigger_delay
3. Send the microcontroller exposure_time for illumination time
4. Send the microcontroller strobe_delay + trigger_delay for strobe
delay so illumination comes on strobe_delay + trigger_delay after the hw
trigger, and stays on for exposure_time

One of the keys in there is calculating trigger_delay using the correct
exposure time (exposure_time for SW trigger, exposure_time +
strobe_delay for HW trigger).

![image_123650291
(3)](https://github.com/user-attachments/assets/5f8f072e-3fa5-4a04-b6e1-8e87237b084d)

There was also a bug in ms -> us conversion that needed fixing.

Tested by: Tested on a system with a toupcam and both hw and software
trigger setup.