MultiPointWorker: protect against in flight triggering (#283)
See title. This is maybe not the exact right fix, but it does work. The
issue is that the QTimer might be running its callback when we stop, so
one image may sneak through. This can then get into the callback at
acquisition start (and so you get an error about capture info not
existing at the start of an acquisition).

Tested by: Running a bunch of acquisitions with live mode enabled to
start, and observing the WARNING prints.