multi process jobs: Add in a job and job runner mechanism for processing images on separate processes (#285)
This adds in a proto multi-process / job running mechanism. Long term,
it'd be much better to use something like dramatiq, huey, rq, celery, or
any of the other well known job queues. But these require some setup
overhead, a broker, and won't get us up and running quickly. Hopefully
the Job and JobRunner mechanism in this PR helps make the swap to that
easier later on.

This is "proto" because it does not allow specifying job pipelines, does
not allow doing anything with the result, and doesn't plumb through
custom jobs yet. All this does is add image saving to the jobs on
separate processes.

Performance improvement numbers to come.

Tested by: Running tests in sim and the hw I have at home. I also tested
adding in the `ThrowImmediatelyJob` and `HangForeverJob` (individually)
as `extra_job_classes` to the `MultiPointWorker`. These behave as
expected so we at least have baseline sane error handling (but no
retries, no recovery from hangs, etc).