bug: toupcam: do not double up on cam exposure time (#296)
We were doubling up the camera strobe + trig delay compensation. This
meant that for longer exposure times, we'd think the camera was ready
for another trigger earlier than it actually was. As a result, we'd get
lost frames (because a trigger too early would be ignored - the camera
was still exposing with the doubled up exposure, but the camera driver
in our code was using the correct frame length).

Tested by: Running an acquisition with >200ms exposure times, and 1x1
binning so that the strobe calc was really long (~147ms). This resulted
in ~500 ms exposures sent to the camera before this fix, but now results
in the correct ~350 ms.