acquisitions: wait for frames when using non-hw triggering (#288)
After discussion with @hongquanli , we decided that for all but HARDWARE
trigger acquisitions, we'll go back to waiting for frames to come in
before turning off illumination. This deals with the non-determinism we
are seeing in the SW triggering across cameras.

Tested by: Running a small 2 channel acqusition with 3 z and 2
timepoints. Mosaic views appear correct. Acquisition image folders
appear correct as well.