contrast autofocus: pull QT out of the base contrast and remove qt from related tests (#331)
I narrowed down our occasionally segfaulting tests to the QT usage in
the contrast autofocus code. We needed to remove the qt usage here
anyway, so this does that and simultaneously fixes the segfaulting test
setup (by separating stubs that no longer need qt, and those that do,
and removing qt usage in tests where not needed).

I think as long as we're using QT without understanding the core problem
that's leading to these segfaults, we're liable to have this return. But
for now, this fixes it.

Tested By: Running on a cephla stage + toupcam system, and making sure
manual contrast autofocus works. Then running an acquisition with
contrast autofocus, and one without, and making sure both work. Unit
tests also now pass without a segfault.