microcontroller: force int as a bandaid, warn on non int arg to pid (#318)
This does 2 things. First, it puts a bandaid on our int<->payload
helpers to make sure they always return an int (since it is assumed they
do in code). Second, for one specific case, it prints a warning if the
user input is integral. It'd be better to error out instead, but this
makes sure not to break existing users that happen to be working. If the
user is giving us a fractional `transitions_per_revolution`, something
is wrong (since a fractional value does not make sense there - we can
only send integers to the micro!). So it would make more sense to make
this a failure!

Tested by: New unit test passes, and existing still pass.