bug fix: format of the coordinates file used for fluidics widget is different from the one in wellplate multipoint widget (#364)
This pull request updates the CSV column naming convention in the
`load_coordinates` method in `software/control/widgets.py` to match a
new standard. The most important change is the switch from capitalized
column names with underscores to lowercase names with spaces and units
in parentheses. This affects both validation and data extraction logic.

**CSV column naming convention update:**

* Changed required columns from `"Region"`, `"X_mm"`, and `"Y_mm"` to
`"region"`, `"x (mm)"`, and `"y (mm)"` for CSV validation and data
extraction, ensuring compatibility with the new CSV format.